+++ test/support/involved_company_create_test_helper.rb
@@ -1,19 +1,31 @@
 require "./test/support/igdb_api_test_helper.rb"
+require "./test/support/twitch_oauth_test_helper.rb"
 
 module InvolvedCompanyCreateTestHelper
   include IgdbApiTestHelper
+  include TwitchOauthTestHelper
 
-  def stub_successful_involved_company_requests(ids, twitch_bearer_token)
-    ids.each do |id|
+  def stub_successful_involved_company_responses
+    igdb_game_data["involved_companies"].each do |id|
       stub_successful_igdb_api_request(
         "involved_companies/#{id}",
         json_mocks("igdb/involved_companies/#{id}.json"),
-        twitch_bearer_token,
+        stubbed_twitch_bearer_token,
       )
     end
   end
 
-  def stub_involved_company_request_failures(ids)
-    ids.each { |id| stub_igdb_api_request_failure("involved_companies/#{id}") }
+  def stub_involved_company_request_failures
+    igdb_game_data["involved_companies"].each do |id|
+      stub_igdb_api_request_failure("involved_companies/#{id}")
+    end
+  end
+
+  def stub_involved_company_responses(with_involved_company_failures)
+    if with_involved_company_failures
+      return stub_involved_company_request_failures
+    end
+
+    stub_successful_involved_company_responses
   end
 end
